+++ src/icons/index.ts
@@ -1,4 +1,5 @@
 export { Icon, type IconProps } from "./Icon";
+export { DatabaseIcon, type DatabaseIconProps } from "./brands";
 export {
   BookmarkIcon,
   CheckIcon,